third_party/blink/web_tests/fast/css/padding-no-renderer-expected.txt

PASS window.getComputedStyle(test0, null).paddingTop is "0px"
PASS window.getComputedStyle(test1, null).paddingTop is "10px"
PASS window.getComputedStyle(test2, null).paddingTop is "10px"
PASS window.getComputedStyle(test3, null).paddingTop is "10px"
PASS window.getComputedStyle(test4, null).paddingTop is "10px"
PASS window.getComputedStyle(test5, null).paddingTop is "10px"
PASS window.getComputedStyle(test6, null).paddingTop is "10%"
PASS window.getComputedStyle(test7, null).paddingTop is "10%"
PASS successfullyParsed is true

TEST COMPLETE
